+++ template-parts/newBlocks/_block_banners.php
@@ -9,7 +9,12 @@
  * @version 1.00
  */
 
+// TOOD: Convert the basic image to use the slider options but only display one image, then remove these fields.
+
 switch ($block['layout']):
+    case "image":
+        include('banners/__image.php');
+        break;
     case "video":
         include('banners/__video.php');
         break;