+++ template-parts/newBlocks/_block_linkBoxes.php
@@ -12,8 +12,8 @@
 $layout = get_sub_field($current . '_layout');
 
 switch ($layout):
-    case "basic":
-        include ('linkedBoxes/__basic.php');
+    case "grid":
+        include ('linkedBoxes/__grid.php');
         break;
     case "title_below":
         include ('linkedBoxes/__title_below.php');