+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.8 – 21 Feb 2018 (1)
+Banners – Added support for sub headings and logos. Added new banner type allowing image overlay.
+Linked Boxes – Add support for linked boxes grid layout (1 Large, 1 Half Height, 2 Half Height/Width).
+
 ## 1.0.7 - 20 Feb 2018 (2) 
 BugFix: Small issues where blocks where loading colours set in the css on new blocks. Also small spacing issues of content block.